ProjectTemplates/scripts/requirements.txt
# 核心依赖包
# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 机器学习
scikit-learn>=1.1.0
scipy>=1.9.0

# 配置文件处理
PyYAML>=6.0
python-dotenv>=0.19.0

# 网络请求
requests>=2.28.0
urllib3>=1.26.0

# 文件处理
openpyxl>=3.0.0
xlrd>=2.0.0

# 日志和调试
colorlog>=6.6.0

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0

# 代码质量
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# 开发工具
jupyter>=1.0.0
ipython>=8.0.0

# 可选依赖（根据需要取消注释）
# Web 开发
# flask>=2.2.0
# fastapi>=0.85.0
# django>=4.1.0

# 数据库
# sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0
# pymongo>=4.2.0

# 异步处理
# asyncio
# aiohttp>=3.8.0

# 图像处理
# Pillow>=9.0.0
# opencv-python>=4.6.0

# 自然语言处理
# nltk>=3.7
# spacy>=3.4.0

# 深度学习
# torch>=1.12.0
# tensorflow>=2.10.0

# 时间处理
# python-dateutil>=2.8.0
# pytz>=2022.1

# 加密和安全
# cryptography>=38.0.0
# bcrypt>=4.0.0
